edited styles +Admin Panel Functionalities

+++ src/pages/admin/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import "./admin.styles.css";
 import { Button } from "@components/button";
 import { BASE_URL } from "@services/apiService";
@@ -12,6 +12,7 @@ export function AdminPage() {
     gender: string;
     phoneNumber: string;
     location: string;
+    url: string;
     email: string;
     password: string;
   }
@@ -21,6 +22,8 @@ export function AdminPage() {
   const [editingTherapist, setEditingTherapist] = useState<Therapist | null>(
     null
   );
+  const editFormRef = useRef<HTMLDivElement>(null);
+
   const [newTherapist, setNewTherapist] = useState({
     firstName: "",
     lastName: "",
@@ -28,6 +31,7 @@ export function AdminPage() {
     gender: "",
     phoneNumber: "",
     location: "",
+    url: "",
     email: "",
     password: "",
   });
@@ -38,18 +42,15 @@ export function AdminPage() {
 
   const fetchTherapists = async () => {
     try {
-      const response = await fetch(
-        `${BASE_URL}/admin`,
-        {
-          method: "GET",
-        }
-      );
+      const response = await fetch(`${BASE_URL}/admin`, {
+        method: "GET",
+      });
 
       if (!response.ok) {
         throw new Error("Failed to fetch therapists");
       }
 
-      const data = await response.json();      
+      const data = await response.json();
       setTherapists(data);
       setOriginalTherapists(data);
     } catch (error) {
@@ -59,20 +60,25 @@ export function AdminPage() {
 
   const addTherapist = async () => {
     try {
-      const response = await fetch(
-        `${BASE_URL}/admin`,
-        {
-          method: "POST",
-          body: JSON.stringify(newTherapist),
+      // Check if any of the fields are empty
+      for (const key in newTherapist) {
+        if (newTherapist[key as keyof typeof newTherapist] === "") {
+          alert("Please fill in all fields");
+          return;
         }
-      );
+      }
+
+      const response = await fetch(`${BASE_URL}/admin`, {
+        method: "POST",
+        body: JSON.stringify(newTherapist),
+      });
 
       if (!response.ok) {
         throw new Error("Failed to add therapist");
       }
 
       const data = await response.json();
-      console.log(data); //log the response
+      console.log(data); // log the response
       fetchTherapists(); // Refresh the therapists list
       setNewTherapist({
         firstName: "",
@@ -81,6 +87,7 @@ export function AdminPage() {
         gender: "",
         phoneNumber: "",
         location: "",
+        url: "",
         email: "",
         password: "",
       }); // Reset the new therapist form
@@ -91,14 +98,18 @@ export function AdminPage() {
 
   const deleteTherapist = async (therapistId: string) => {
     try {
-      const response = await fetch(
-        `${BASE_URL}/admin`,
-        {
-          method: "DELETE",
-          credentials: "include",
-          body: JSON.stringify({ therapistId }),
-        }
+      const confirmDelete = window.confirm(
+        "Are you sure you want to delete this therapist?"
       );
+      if (!confirmDelete) {
+        return; // User clicked cancel, do nothing
+      }
+
+      const response = await fetch(`${BASE_URL}/admin`, {
+        method: "DELETE",
+        credentials: "include",
+        body: JSON.stringify({ therapistId }),
+      });
 
       if (!response.ok) {
         throw new Error("Failed to delete therapist");
@@ -111,27 +122,26 @@ export function AdminPage() {
       console.error(error);
     }
   };
-  function searchTable(searchTerm : string) {
-    
-    setTherapists(originalTherapists.filter( x => x.firstName.includes(searchTerm)));
-    
+  function searchTable(searchTerm: string) {
+    setTherapists(
+      originalTherapists.filter((x) => x.firstName.includes(searchTerm))
+    );
   }
   const editTherapist = async () => {
     try {
       if (editingTherapist) {
         // Send the updated therapist data to the server and update state
-        const response = await fetch(
-          `${BASE_URL}/admin`,
-          {
-            method: "PUT",
-            credentials: "include",
-            body: JSON.stringify(editingTherapist),
-          }
-        );
-          
+        const response = await fetch(`${BASE_URL}/admin`, {
+          method: "PUT",
+          credentials: "include",
+          body: JSON.stringify(editingTherapist),
+        });
+
         if (response.ok) {
           const updatedTherapists = therapists.map((therapist) =>
-            therapist._id === editingTherapist._id ? editingTherapist : therapist
+            therapist._id === editingTherapist._id
+              ? editingTherapist
+              : therapist
           );
           setTherapists(updatedTherapists);
           setOriginalTherapists(updatedTherapists);
@@ -146,18 +156,29 @@ export function AdminPage() {
   };
   const handleEditClick = (therapist: Therapist) => {
     setEditingTherapist(therapist);
-  };
+    const offset = 100; // Adjust the offset value as needed
 
+    if (editFormRef.current) {
+      const { top } = editFormRef.current.getBoundingClientRect();
+      window.scrollTo({
+        top: window.pageYOffset + top - offset,
+        behavior: "smooth",
+      });
+    }
+  };
   const handleCancelEdit = () => {
     setEditingTherapist(null);
   };
+
   return (
     <div>
-    <input
-      type="text"
-      placeholder="Search"
-      onChange={(e) => searchTable(e.currentTarget.value)}>
-    </input>
+      <div className="search-bar">
+        <input
+          type="text"
+          placeholder="Search"
+          onChange={(e) => searchTable(e.currentTarget.value)}
+        ></input>
+      </div>
       {/* Therapist Table */}
       <table>
         <thead>
@@ -168,6 +189,7 @@ export function AdminPage() {
             <th>Gender</th>
             <th>Phone Number</th>
             <th>Location</th>
+            <th>Profile Picture Link</th>
             <th>Email</th>
             <th>Password</th>
             <th>Actions</th>
@@ -182,6 +204,7 @@ export function AdminPage() {
               <td>{therapist.gender}</td>
               <td>{therapist.phoneNumber}</td>
               <td>{therapist.location}</td>
+              <td>{therapist.url}</td>
               <td>{therapist.email}</td>
               <td>{therapist.password}</td>
               <td>
@@ -202,10 +225,11 @@ export function AdminPage() {
         <input
           className="admin-input"
           type="text"
-          placeholder="Fist Name"
+          placeholder="First Name"
           onChange={(e) =>
             setNewTherapist({ ...newTherapist, firstName: e.target.value })
           }
+          required // Add the required attribute
         />
 
         <label className="labels">Last Name:</label>
@@ -216,6 +240,7 @@ export function AdminPage() {
           onChange={(e) =>
             setNewTherapist({ ...newTherapist, lastName: e.target.value })
           }
+          required // Add the required attribute
         />
 
         <label className="labels">Date of Birth:</label>
@@ -226,6 +251,7 @@ export function AdminPage() {
           onChange={(e) =>
             setNewTherapist({ ...newTherapist, dob: e.target.value })
           }
+          required // Add the required attribute
         />
 
         <label className="labels">Gender:</label>
@@ -235,6 +261,7 @@ export function AdminPage() {
           onChange={(e) =>
             setNewTherapist({ ...newTherapist, gender: e.target.value })
           }
+          required // Add the required attribute
         >
           <option value="">Select Gender</option>
           <option value="Male">Male</option>
@@ -249,6 +276,7 @@ export function AdminPage() {
           onChange={(e) =>
             setNewTherapist({ ...newTherapist, phoneNumber: e.target.value })
           }
+          required // Add the required attribute
         />
 
         <label className="labels">Location:</label>
@@ -259,6 +287,18 @@ export function AdminPage() {
           onChange={(e) =>
             setNewTherapist({ ...newTherapist, location: e.target.value })
           }
+          required // Add the required attribute
+        />
+
+        <label className="labels">Profile Picture:</label>
+        <input
+          className="admin-input"
+          type="text"
+          placeholder="Profile Picture"
+          onChange={(e) =>
+            setNewTherapist({ ...newTherapist, url: e.target.value })
+          }
+          required // Add the required attribute
         />
 
         <label className="labels">Email:</label>
@@ -269,6 +309,7 @@ export function AdminPage() {
           onChange={(e) =>
             setNewTherapist({ ...newTherapist, email: e.target.value })
           }
+          required // Add the required attribute
         />
 
         <label className="labels">Password:</label>
@@ -279,6 +320,7 @@ export function AdminPage() {
           onChange={(e) =>
             setNewTherapist({ ...newTherapist, password: e.target.value })
           }
+          required // Add the required attribute
         />
       </div>
 
@@ -302,6 +344,7 @@ export function AdminPage() {
                   firstName: e.target.value,
                 })
               }
+              required // Add the required attribute
             />
             <label className="labels">Last Name:</label>
             <input
@@ -314,6 +357,7 @@ export function AdminPage() {
                   lastName: e.target.value,
                 })
               }
+              required // Add the required attribute
             />
 
             <label className="labels">Date of Birth:</label>
@@ -327,6 +371,7 @@ export function AdminPage() {
                   dob: e.target.value,
                 })
               }
+              required // Add the required attribute
             />
 
             <label className="labels">Gender:</label>
@@ -339,6 +384,7 @@ export function AdminPage() {
                   gender: e.target.value,
                 })
               }
+              required // Add the required attribute
             >
               <option value="">Select Gender</option>
               <option value="Male">Male</option>
@@ -356,6 +402,7 @@ export function AdminPage() {
                   phoneNumber: e.target.value,
                 })
               }
+              required // Add the required attribute
             />
 
             <label className="labels">Location:</label>
@@ -369,6 +416,21 @@ export function AdminPage() {
                   location: e.target.value,
                 })
               }
+              required // Add the required attribute
+            />
+
+            <label className="labels">Profile Picture:</label>
+            <input
+              className="admin-input"
+              type="text"
+              value={editingTherapist.url}
+              onChange={(e) =>
+                setEditingTherapist({
+                  ...editingTherapist,
+                  url: e.target.value,
+                })
+              }
+              required // Add the required attribute
             />
 
             <label className="labels">Email:</label>
@@ -382,27 +444,30 @@ export function AdminPage() {
                   email: e.target.value,
                 })
               }
+              required // Add the required attribute
             />
-
-            <label className="labels">Password:</label>
-            <input
-              className="admin-input"
-              type="text"
-              value={editingTherapist.password}
-              onChange={(e) =>
-                setEditingTherapist({
-                  ...editingTherapist,
-                  password: e.target.value,
-                })
-              }
-            />
+            <div ref={editFormRef}>
+              <label className="labels">Password:</label>
+              <input
+                className="admin-input"
+                type="text"
+                value={editingTherapist.password}
+                onChange={(e) =>
+                  setEditingTherapist({
+                    ...editingTherapist,
+                    password: e.target.value,
+                  })
+                }
+                required // Add the required attribute
+              />
+            </div>
+            <Button className="admin-edit" onClick={handleCancelEdit}>
+              Cancel
+            </Button>
+            <Button className="admin-edit" onClick={editTherapist}>
+              Save
+            </Button>
           </div>
-          <Button className="admin-edit" onClick={handleCancelEdit}>
-            Cancel
-          </Button>
-          <Button className="admin-edit" onClick={editTherapist}>
-            Save
-          </Button>
         </div>
       )}
     </div>
